% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-template.R
\name{analysis_template}
\alias{analysis_template}
\title{Use Analysis Template}
\usage{
analysis_template(analysis, open = rlang::is_interactive())
}
\arguments{
\item{analysis}{A string of name of the analysis.}

\item{open}{Whether or not to open the script interactively; defaults to value returned by \code{rlang::is_interactive}.}
}
\description{
Creates a templated R file in the root directory of the project. The name of
the file takes the value of the \code{analysis} argument.
}
\examples{
\dontrun{
# This will create a file called "abundance.R" in the root directory.
analysis_template('abundance')
}
}
