% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro2jags-package.R
\docType{package}
\name{intro2jags-package}
\alias{intro2jags}
\alias{intro2jags-package}
\title{intro2jags: Introduction to Coding Bayesian Models Course Package}
\description{
Loads useful packages, datasets, and functions for the CMI Introduction to Coding Bayesian Models Course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/intro2jags/}
  \item \url{https://github.com/poissonconsulting/intro2jags}
  \item Report bugs at \url{https://github.com/poissonconsulting/intro2jags/issues}
}

}
\author{
\strong{Maintainer}: Nicole Hill \email{nicole@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7623-2153}{ORCID})

}
\keyword{internal}
