% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{analysis}{A string of name of the analysis.}

\item{open}{Whether or not to open the script interactively; defaults to value returned by \code{rlang::is_interactive}.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.
}
\keyword{internal}
