% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-starfish.R
\docType{data}
\name{starfish}
\alias{starfish}
\title{Counts of starfish in square metre quadrats.}
\format{
A tibble with 150 rows and 4 variables:
\describe{
\item{count}{an integer denoting the number of starfish observed in a 1 metre^2 quadrat}
\item{temp}{a number denoting the water temperature (degrees Celcius)}
\item{site}{a factor denoting the site name (Site A, Site B, or Site C)}
\item{year}{an integer denoting the study year (2011-2020)}
}
}
\usage{
starfish
}
\description{
Includes counts of starfish, temperature measurements, and the site and year
of the observation.
This is a simulated dataset.
}
\keyword{datasets}
