% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-climate.R
\docType{data}
\name{climate}
\alias{climate}
\title{Temperature anomaly measurements.}
\format{
A tibble with 100 rows and 3 variables:
\describe{
\item{temp_anaomaly}{a number denoting the deviation from a baseline temperature}
\item{elev}{a number denoting the elevation (metres)}
\item{dist_coast}{a number denoting the distance from the coast (kilometres)}
}
}
\usage{
climate
}
\description{
Includes temperature anomalys from 100 locations, with measurements of the
elevation and distance from the coast for each observation.
This is a simulated dataset.
}
\keyword{datasets}
